chore(core): move request function types from react-components to core package
