+++ packages/react-components/src/components/resource-explorers/explorers/asset-explorer/types.ts
@@ -8,7 +8,7 @@ import type {
   ExecuteQuery,
   ListAssets,
   ListAssociatedAssets,
-} from '../../types/request-fn';
+} from '@iot-app-kit/core';
 import type { CommonResourceExplorerProps } from '../../types/resource-explorer';
 import type { AssetResource } from '../../types/resources';
